in-browser db
https://news.ycombinator.com/item?id=24806344

So a storage abstraction layer on top of IndexDB with a proprietery API.
...that synchronizes with a server-side abstraction layer on top of MongoDB. This is essentially the Meteor architecture.

Synchronization is non-trivial and the storage abstraction options come down to NoSQL JSON document stores, GraphQL, and SQL. Browsers support IndexedDB while native iOS/Android support SQLite. I’m not aware of any decent solution that synchronizes a relational database with client-side SQLite.

I too would prefer a SQLite Sync solution that just works and SQLite in the browser available from Service Workers but it is what it is.